requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0

# Agent framework (暂时注释，Sprint 2实现)
# langgraph>=0.0.40
# langchain>=0.1.0
# langchain-community>=0.0.20

# LLM integration (暂时注释，Sprint 2实现)
# openai>=1.0.0

# Web framework
streamlit>=1.28.0
requests>=2.31.0

# Task scheduling (暂时注释，Sprint 2实现)
# apscheduler>=3.10.0

# Data processing
pandas>=2.0.0

# Logging and monitoring
structlog>=23.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Testing utilities
faker>=19.0.0
factory-boy>=3.3.0
responses>=0.23.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
